refactor: separate clients into their own package
These API client will be used from both the frontend and the data upload
functionality.
