+++ packages/clients/package.json
@@ -1,0 +1,16 @@
+{
+  "private": "true",
+  "name": "@csfin/clients",
+  "version": "0.0.1",
+  "main": "build/index.js",
+  "types": "build/index.d.ts",
+  "files": [
+    "build"
+  ],
+  "scripts": {
+    "build": "tsc --build --verbose",
+    "lint": "eslint --color --ext .ts --ext .tsx src/",
+    "lint:fix": "eslint --color --fix --ext .ts --ext .tsx src/"
+  },
+  "dependencies": {}
+}
