+++ packages/clients/tsconfig.json
@@ -5,10 +5,5 @@
     "outDir": "build",
     "module": "CommonJS"
   },
-  "references": [
-    { "path": "../../packages/clients" },
-    { "path": "../../packages/codi" },
-    { "path": "../../packages/core" }
-  ],
   "exclude": ["node_modules", "build"]
 }
